Analog handles now return 32-bit darta

+++ src/CommandProcessing/ScanningSensorHandler.h
@@ -16,7 +16,7 @@ namespace ScanningSensorHandler
 {
 	void Init() noexcept;
 	bool IsPresent() noexcept;
-	uint16_t GetReading() noexcept;
+	uint32_t GetReading() noexcept;
 	void AppendDiagnostics(const StringRef& reply) noexcept;
 }
